ingestor/requirements.txt
confluent-kafka>=2.14.0
pandas==2.2.3
pyarrow==17.0.0
redis==5.1.1
python-dotenv==1.0.1
